shade the undertow + xnio classes
We had problems with using this in a test for a spring-boot application,
because Spring-Boot's autoconfiguration decided to use undertow as its
web-server (in parallel with Tomcat) when that was in the class path.

By renaming all the class names, this is avoided.

We had to re-add one service loader hint file, as the class name changed.